+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [#176](https://github.com/hyperf-cloud/hyperf/pull/176) Fixed TypeError: Return value of LengthAwarePaginator::nextPageUrl() must be of the type string or null, none returned.
 - [#188](https://github.com/hyperf-cloud/hyperf/pull/188) Fixed proxy of guzzle client does not work expected.
 - [#211](https://github.com/hyperf-cloud/hyperf/pull/211) Fixed rpc clients will be replaced by the latest one. 
+- [#212](https://github.com/hyperf-cloud/hyperf/pull/212) Fixed config `ssl_key` and `cert` of guzzle client does not work expected.
 
 # v1.0.4 - 2019-07-08